Line level comment box not displayed for file type html or js
The issue is with setting the Annotations, ACE syntax validation has its own annotations which removes/mendles with the custom annotations for our code ocean comments. This is a known issue refer the below github issue.

securingsincity/react-ace#483

To overcome this I added a flag to our custom annotations and seperate it from the ACE annotations and then push them together as annotations by having an event Listener on change of annotation refer to the below code .